+++ src/main/java/nu/olivertwistor/currencymgr/App.java
@@ -1,6 +1,5 @@
 package nu.olivertwistor.currencymgr;
 
-import nu.olivertwistor.currencymgr.ui.GUI;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -13,7 +12,7 @@ import java.lang.reflect.InvocationTargetException;
  *
  * @since 0.1.0
  */
-@SuppressWarnings({"HardCodedStringLiteral", "ClassOnlyUsedInOnePackage"})
+@SuppressWarnings("HardCodedStringLiteral") //todo make translatable
 final class App
 {
     private static final Logger LOG = LogManager.getLogger();