+++ src/main/java/nu/olivertwistor/currencymgr/actions/SaveAsFileAction.java
@@ -1,6 +1,6 @@
-package nu.olivertwistor.currencymgr.ui.actions;
+package nu.olivertwistor.currencymgr.actions;
 
-import nu.olivertwistor.currencymgr.ui.GUI;
+import nu.olivertwistor.currencymgr.GUI;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -14,7 +14,7 @@ import java.awt.event.KeyEvent;
  *
  * @since 0.1.0
  */
-@SuppressWarnings("HardCodedStringLiteral")
+@SuppressWarnings("HardCodedStringLiteral") //todo make translatable
 public final class SaveAsFileAction extends AbstractDialogSaveAction
 {
     private static final Logger LOG = LogManager.getLogger();