+++ src/main/java/nu/olivertwistor/currencymgr/actions/SaveFileAction.java
@@ -1,6 +1,6 @@
-package nu.olivertwistor.currencymgr.ui.actions;
+package nu.olivertwistor.currencymgr.actions;
 
-import nu.olivertwistor.currencymgr.ui.GUI;
+import nu.olivertwistor.currencymgr.GUI;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -18,11 +18,11 @@ import java.awt.event.KeyEvent;
  * @since 0.1.0
  */
 @SuppressWarnings({"HardCodedStringLiteral", "ClassOnlyUsedInOnePackage"})
+//todo make translatable
 public final class SaveFileAction extends AbstractAction
 {
     private static final Logger LOG = LogManager.getLogger();
 
-    @SuppressWarnings("unused")
     private final GUI gui;
 
     /**