+++ src/main/java/nu/olivertwistor/currencymgr/actions/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains all of the Swing actions present in the app.
+ *
+ * @since 0.1.0
+ * @since x Moved this package up one level in the hierarchy. //todo correct version
+ */
+package nu.olivertwistor.currencymgr.actions;