+++ src/main/java/nu/olivertwistor/currencymgr/ui/actions/package-info.java
@@ -1,6 +1,0 @@
-/**
- * This package contains all of the Swing actions present in the app.
- *
- * @since 0.1.0
- */
-package nu.olivertwistor.currencymgr.ui.actions;