+++ src/main/resources/db/0/create-table-currency.sql
@@ -1,0 +1,6 @@
+CREATE TABLE IF NOT EXISTS currency
+(
+    id     INTEGER PRIMARY KEY,
+    ticker TEXT    NOT NULL,
+    name   TEXT    DEFAULT ''
+);