+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1] and this project adheres to [Semantic Versioning 2.0.0][2].
 
 ## Unreleased
+* Added
+  * Database upgrade scripts for db_version and currency tables.
+  * Ability to perform database upgrades.
+  * Currency data model.
 
 ## [0.1.0](https://github.com/olivertwistor/currency-manager/releases/tag/v0.1.0)
 * Added